+++ tests/pytools/lecpy/test_lecpy.py
@@ -6,35 +6,13 @@ from tempfile import TemporaryDirectory
 import numpy as np
 import pandas as pd
 import pytest
-from oasislmf.pytools.common.data import occurrence_dtype, summary_stream_index_dtype
+from oasislmf.pytools.common.data import occurrence_dtype
 from oasislmf.pytools.common.event_stream import SUMMARY_STREAM_ID, stream_info_to_bytes
 from oasislmf.pytools.lec.data import EPT_dtype, PSEPT_dtype
 from oasislmf.pytools.lec.manager import main
+from tests.pytools.utils import make_idx_from_bin
 
 TESTS_ASSETS_DIR = Path(__file__).parent.parent.parent.joinpath("assets").joinpath("test_lecpy")
-
-
-def make_idx_from_bin(bin_path: Path, idx_path: Path) -> None:
-    """Scan a summary .bin and write a paired .idx recording each event block's byte offset.
-
-    If the bin contains no data blocks (header only), creates a 0-byte idx to match
-    summarypy --low-memory behaviour for partitions that received no events.
-    """
-    raw = np.fromfile(str(bin_path), dtype=np.int32)
-    pos = 3  # skip 3-int stream header (stream_type, sample_size, summary_set_id)
-    entries = []
-    while pos < len(raw):
-        byte_offset = pos * 4
-        summary_id = int(raw[pos + 1])
-        pos += 3  # event_id, summary_id, expval
-        while pos < len(raw) and raw[pos] != 0:
-            pos += 2  # sidx + loss pair (any non-zero sidx, including special negatives)
-        pos += 2  # terminating (sidx=0, loss=0.0)
-        entries.append((summary_id, byte_offset))
-    if entries:
-        np.array(entries, dtype=summary_stream_index_dtype).tofile(str(idx_path))
-    else:
-        idx_path.touch()  # empty partition → 0-byte idx
 
 
 def case_runner(sub_folder, test_name, out_ext="csv", use_return_period=False):
